requirements.txt
python-dotenv>=1.0.0
discord.py>=2.3.2
Flask>=3.0.0
yt-dlp>=2023.11.16
mutagen>=1.47.0
python-magic>=0.4.27
ffmpeg-python>=0.2.0
PyNaCl>=1.5.0
pytube>=15.0.0
requests>=2.31.0
aiohttp>=3.9.1
websockets>=12.0